+++ src/pyro/exception.clj
@@ -1,4 +1,0 @@
-(ns pyro.exception
-  (:require [clojure.main :as main]
-            [clojure.repl :as repl]
-            [clojure.stacktrace :as st]))