+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.1.0
- * Drop all stacktrace frames once we hit `clojure.main/repl/read-eval-print`.
+ * Source code reader that highlights the relevant line in question
+ * Ability to drop all stacktrace frames once we hit `clojure.main/repl/read-eval-print`.